Reorg and rename mtermvectors
